feat: 새로고침 감지 함수  hook으로 구현
