+++ src/app/products/new/page.tsx
@@ -24,23 +24,11 @@ function RegisterProduct() {
     resolver: zodResolver(ProductSchema)
   });
 
+  const setActive = useBeforeUnload();
 
   const onSubmit: SubmitHandler<RegisterProduct> = (data) => {
     console.log(data);
   };
-
-  /** 새로고침 감지 */
-  useEffect(() => {
-    const handleBeforeUnload = (event: BeforeUnloadEvent) => {
-      event.preventDefault();
-    };
-
-    window.addEventListener("beforeunload", handleBeforeUnload);
-
-    return () => {
-      window.removeEventListener("beforeunload", handleBeforeUnload);
-    };
-  }, []);
 
   return (
     <main>
